requirements.txt
streamlit
python-dotenv
youtube-transcript-api
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
faiss-cpu
yt-dlp
